+++ src/functions.ts
@@ -6,7 +6,10 @@ import { select } from "./functions/select"
 import { device } from "./functions/device"
 import { stack } from "./functions/stack"
 
-export type icFunction = (env: Environment, data: (string | number)[]) => void | Promise<void> | Error[] | Promise<Error[]>
+export type icFunction = (
+	env: Environment,
+	data: (string | number)[],
+) => void | Promise<void> | Error[] | Promise<Error[]>
 export type icCondition = (env: Environment, data: (string | number)[]) => boolean
 
 export const functions: { [key: string]: icFunction } = {