+++ src/tests/main.test.ts
@@ -5,7 +5,7 @@ import data from "./data/data.json"
 describe("main", () => {
 	test("functions", () => {
 		let functionNames: string[] = []
-		Object.entries(data).map(function([key, value]) {
+		Object.entries(data).map(function ([key, value]) {
 			if (value.type === "Function") {
 				if (["debug", "stack", "return"].includes(key)) {
 					return