+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 [29-04-2021 23:9]
 
-- [new]: add support `sb`,`lr`,`lb`
-- [new]: add special function: `#d0-5 [hash]`
+-   [new]: add support `sb`,`lr`,`lb`
+-   [new]: add special function: `#d0-5 [hash]`
 
 [21-04-2021 15:52]
 
-- [fix]: add RatioNitrousOxide to Device
+-   [fix]: add RatioNitrousOxide to Device